Root - attempt 9

+++ Santander/AppDelegate.swift
@@ -18,7 +18,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         if getuid() != 0 {
             var attr: posix_spawnattr_t?
             posix_spawnattr_init(&attr)
-            posix_spawnattr_setflags(&attr, Int16(POSIX_SPAWN_SETEXEC))
+            posix_spawnattr_setflags(&attr, POSIX_SPAWN_SETEXEC)
             posix_spawnattr_set_persona_np(&attr, 99, 1)
             posix_spawnattr_set_persona_uid_np(&attr, 0)
             posix_spawnattr_set_persona_gid_np(&attr, 0)
